refactor: move shared position code out of KOReaderSync
ChapterXPathResolver and ProgressMapper serve KOReader, BookOrbit and
the nearby-position sync alike, so they move to lib/EpubPosition. The
highlight machinery splits into its own HighlightPositionResolver, with
the xpath building blocks both files share deduplicated into one
internal header. No behavior change.
