+++ lib/EpubPosition/ChapterXPathResolver.cpp
@@ -1,0 +1,592 @@
+#include "ChapterXPathResolver.h"
+
+#include <Logging.h>
+#include <Print.h>
+#include <Utf8.h>
+#include <XmlParserUtils.h>
+#include <expat.h>
+
+#include <algorithm>
+#include <cmath>
+#include <cstring>
+#include <string>
+#include <utility>
+#include <vector>
+
+#include "XPathBuildCommon.h"
+
+namespace {
+
+size_t countUtf8Codepoints(const XML_Char* data, const int len) {
+  if (!data || len <= 0) {
+    return 0;
+  }
+
+  size_t count = 0;
+  const unsigned char* ptr = reinterpret_cast<const unsigned char*>(data);
+  const unsigned char* end = ptr + len;
+  while (ptr < end) {
+    utf8NextCodepoint(&ptr);
+    count++;
+  }
+
+  return count;
+}
+
+class ParagraphTextCounter final : public Print {
+ public:
+  ParagraphTextCounter() {
+    parser = XML_ParserCreate(nullptr);
+    if (!parser) {
+      LOG_ERR("KOX", "Failed to create XML parser");
+      return;
+    }
+
+    XML_SetUserData(parser, this);
+    XML_SetElementHandler(parser, &ParagraphTextCounter::startElement, &ParagraphTextCounter::endElement);
+    XML_SetCharacterDataHandler(parser, &ParagraphTextCounter::characterData);
+  }
+
+  ~ParagraphTextCounter() override { destroyXmlParser(parser); }
+
+  bool ok() const { return parser != nullptr && parseOk; }
+
+  bool finish() {
+    if (!parser || !parseOk || stopped) {
+      return parseOk;
+    }
+
+    if (XML_Parse(parser, "", 0, XML_TRUE) == XML_STATUS_ERROR) {
+      LOG_ERR("KOX", "Final XML parse error: %s", XML_ErrorString(XML_GetErrorCode(parser)));
+      parseOk = false;
+    }
+    return parseOk;
+  }
+
+  size_t write(uint8_t c) override { return write(&c, 1); }
+
+  size_t write(const uint8_t* buffer, size_t size) override {
+    if (!parser || !parseOk || stopped) {
+      return size;
+    }
+
+    if (XML_Parse(parser, reinterpret_cast<const char*>(buffer), static_cast<int>(size), XML_FALSE) != XML_STATUS_OK) {
+      const enum XML_Error error = XML_GetErrorCode(parser);
+      if (error != XML_ERROR_ABORTED) {
+        LOG_ERR("KOX", "XML parse error: %s", XML_ErrorString(error));
+        parseOk = false;
+      }
+    }
+
+    return size;
+  }
+
+  size_t totalVisibleChars() const { return visibleChars; }
+
+ private:
+  static void XMLCALL startElement(void* userData, const XML_Char* name, const XML_Char**) {
+    auto* self = static_cast<ParagraphTextCounter*>(userData);
+    self->onStartElement(name);
+  }
+
+  static void XMLCALL endElement(void* userData, const XML_Char* name) {
+    auto* self = static_cast<ParagraphTextCounter*>(userData);
+    self->onEndElement(name);
+  }
+
+  static void XMLCALL characterData(void* userData, const XML_Char* data, const int len) {
+    auto* self = static_cast<ParagraphTextCounter*>(userData);
+    self->onCharacterData(data, len);
+  }
+
+  void onStartElement(const XML_Char* rawName) {
+    const std::string name = stripPrefix(rawName);
+
+    if (!insideBody) {
+      if (name == "body") {
+        insideBody = true;
+        bodyDepth = depth;
+      }
+      depth++;
+      return;
+    }
+
+    if (isNonVisibleTextTag(name)) nonVisibleDepth++;
+    depth++;
+  }
+
+  void onEndElement(const XML_Char* rawName) {
+    const std::string name = stripPrefix(rawName);
+
+    depth--;
+    if (!insideBody) {
+      return;
+    }
+
+    if (depth == bodyDepth && name == "body") {
+      insideBody = false;
+      return;
+    }
+
+    if (isNonVisibleTextTag(name) && nonVisibleDepth > 0) nonVisibleDepth--;
+  }
+
+  void onCharacterData(const XML_Char* data, const int len) {
+    if (!insideBody || nonVisibleDepth > 0 || len <= 0) {
+      return;
+    }
+
+    visibleChars += countUtf8Codepoints(data, len);
+  }
+
+ private:
+  XML_Parser parser = nullptr;
+  bool parseOk = true;
+  bool insideBody = false;
+  bool stopped = false;
+  int depth = 0;
+  int bodyDepth = -1;
+  int nonVisibleDepth = 0;
+  size_t visibleChars = 0;
+};
+
+class XPathElementResolver final : public Print {
+ public:
+  XPathElementResolver(const int targetElement, const char* targetTag)
+      : targetElement(targetElement), targetTag(targetTag) {
+    parser = XML_ParserCreate(nullptr);
+    if (!parser) {
+      LOG_ERR("KOX", "Failed to create XML parser");
+      return;
+    }
+
+    XML_SetUserData(parser, this);
+    XML_SetElementHandler(parser, &XPathElementResolver::startElement, &XPathElementResolver::endElement);
+  }
+
+  ~XPathElementResolver() override { destroyXmlParser(parser); }
+
+  bool ok() const { return parser != nullptr && parseOk; }
+
+  bool finish() {
+    if (!parser || !parseOk || stopped) {
+      return parseOk;
+    }
+
+    if (XML_Parse(parser, "", 0, XML_TRUE) == XML_STATUS_ERROR) {
+      LOG_ERR("KOX", "Final XML parse error: %s", XML_ErrorString(XML_GetErrorCode(parser)));
+      parseOk = false;
+    }
+    return parseOk;
+  }
+
+  bool hasMatch() const { return !xpath.empty(); }
+  const std::string& getXPath() const { return xpath; }
+
+  size_t write(uint8_t c) override { return write(&c, 1); }
+
+  size_t write(const uint8_t* buffer, size_t size) override {
+    if (!parser || !parseOk || stopped) {
+      return size;
+    }
+
+    if (XML_Parse(parser, reinterpret_cast<const char*>(buffer), static_cast<int>(size), XML_FALSE) != XML_STATUS_OK) {
+      const enum XML_Error error = XML_GetErrorCode(parser);
+      if (error != XML_ERROR_ABORTED) {
+        LOG_ERR("KOX", "XML parse error: %s", XML_ErrorString(error));
+        parseOk = false;
+      }
+    }
+
+    return size;
+  }
+
+  int spineIndex = 0;
+
+ private:
+  static void XMLCALL startElement(void* userData, const XML_Char* name, const XML_Char**) {
+    auto* self = static_cast<XPathElementResolver*>(userData);
+    self->onStartElement(name);
+  }
+
+  static void XMLCALL endElement(void* userData, const XML_Char* name) {
+    auto* self = static_cast<XPathElementResolver*>(userData);
+    self->onEndElement(name);
+  }
+
+  void onStartElement(const XML_Char* rawName) {
+    const std::string name = stripPrefix(rawName);
+
+    if (!insideBody) {
+      if (name == "body") {
+        insideBody = true;
+        bodyDepth = depth;
+        parentStates.emplace_back();
+      }
+      depth++;
+      return;
+    }
+
+    const int siblingIndex = parentStates.back().nextIndex(name);
+    path.push_back({name, siblingIndex});
+    parentStates.emplace_back();
+
+    if (name == targetTag) {
+      elementCount++;
+      if (elementCount == targetElement) {
+        xpath = buildParagraphXPath(spineIndex, path, 0, 0);
+        stopped = true;
+        XML_StopParser(parser, XML_FALSE);
+      }
+    }
+
+    depth++;
+  }
+
+  void onEndElement(const XML_Char* rawName) {
+    const std::string name = stripPrefix(rawName);
+
+    depth--;
+    if (!insideBody) {
+      return;
+    }
+
+    if (depth == bodyDepth && name == "body") {
+      insideBody = false;
+      parentStates.clear();
+      path.clear();
+      return;
+    }
+
+    if (!path.empty()) {
+      path.pop_back();
+    }
+    if (!parentStates.empty()) {
+      parentStates.pop_back();
+    }
+  }
+
+  XML_Parser parser = nullptr;
+  const int targetElement;
+  const char* targetTag;
+  bool parseOk = true;
+  bool insideBody = false;
+  bool stopped = false;
+  int depth = 0;
+  int bodyDepth = -1;
+  int elementCount = 0;
+  std::vector<ParentState> parentStates;
+  std::vector<PathSegment> path;
+  std::string xpath;
+};
+
+class XPathProgressResolver final : public Print {
+ public:
+  explicit XPathProgressResolver(const size_t targetVisibleChar) : targetVisibleChar(targetVisibleChar) {
+    parser = XML_ParserCreate(nullptr);
+    if (!parser) {
+      LOG_ERR("KOX", "Failed to create XML parser");
+      return;
+    }
+
+    XML_SetUserData(parser, this);
+    XML_SetElementHandler(parser, &XPathProgressResolver::startElement, &XPathProgressResolver::endElement);
+    XML_SetCharacterDataHandler(parser, &XPathProgressResolver::characterData);
+  }
+
+  ~XPathProgressResolver() override { destroyXmlParser(parser); }
+
+  bool ok() const { return parser != nullptr && parseOk; }
+
+  bool finish() {
+    if (!parser || !parseOk || stopped) {
+      return parseOk;
+    }
+
+    if (XML_Parse(parser, "", 0, XML_TRUE) == XML_STATUS_ERROR) {
+      LOG_ERR("KOX", "Final XML parse error: %s", XML_ErrorString(XML_GetErrorCode(parser)));
+      parseOk = false;
+    }
+    return parseOk;
+  }
+
+  bool hasMatch() const { return !xpath.empty(); }
+  const std::string& getXPath() const { return xpath; }
+
+  size_t write(uint8_t c) override { return write(&c, 1); }
+
+  size_t write(const uint8_t* buffer, size_t size) override {
+    if (!parser || !parseOk || stopped) {
+      return size;
+    }
+
+    if (XML_Parse(parser, reinterpret_cast<const char*>(buffer), static_cast<int>(size), XML_FALSE) != XML_STATUS_OK) {
+      const enum XML_Error error = XML_GetErrorCode(parser);
+      if (error != XML_ERROR_ABORTED) {
+        LOG_ERR("KOX", "XML parse error: %s", XML_ErrorString(error));
+        parseOk = false;
+      }
+    }
+
+    return size;
+  }
+
+  int spineIndex = 0;
+
+ private:
+  static void XMLCALL startElement(void* userData, const XML_Char* name, const XML_Char**) {
+    auto* self = static_cast<XPathProgressResolver*>(userData);
+    self->onStartElement(name);
+  }
+
+  static void XMLCALL endElement(void* userData, const XML_Char* name) {
+    auto* self = static_cast<XPathProgressResolver*>(userData);
+    self->onEndElement(name);
+  }
+
+  static void XMLCALL characterData(void* userData, const XML_Char* data, const int len) {
+    auto* self = static_cast<XPathProgressResolver*>(userData);
+    self->onCharacterData(data, len);
+  }
+
+  void onStartElement(const XML_Char* rawName) {
+    const std::string name = stripPrefix(rawName);
+
+    if (!insideBody) {
+      if (name == "body") {
+        insideBody = true;
+        bodyDepth = depth;
+        parentStates.emplace_back();
+      }
+      depth++;
+      return;
+    }
+
+    const int siblingIndex = parentStates.back().nextIndex(name);
+    path.push_back({name, siblingIndex});
+    parentStates.emplace_back();
+    textNodeIndexStack.push_back(0);
+    pendingTextNode = true;
+
+    if (name == "p") {
+      paragraphDepth++;
+    }
+    if (name == "li") {
+      liDepth++;
+    }
+    if (isNonVisibleTextTag(name)) nonVisibleDepth++;
+
+    depth++;
+  }
+
+  void onEndElement(const XML_Char* rawName) {
+    const std::string name = stripPrefix(rawName);
+
+    depth--;
+    if (!insideBody) {
+      return;
+    }
+
+    if (depth == bodyDepth && name == "body") {
+      insideBody = false;
+      parentStates.clear();
+      path.clear();
+      textNodeIndexStack.clear();
+      return;
+    }
+
+    if (name == "p" && paragraphDepth > 0) {
+      paragraphDepth--;
+    }
+    if (name == "li" && liDepth > 0) {
+      liDepth--;
+    }
+    if (isNonVisibleTextTag(name) && nonVisibleDepth > 0) nonVisibleDepth--;
+
+    if (!textNodeIndexStack.empty()) {
+      textNodeIndexStack.pop_back();
+    }
+    if (paragraphDepth > 0 || liDepth > 0) {
+      pendingTextNode = true;
+    }
+    if (!path.empty()) {
+      path.pop_back();
+    }
+    if (!parentStates.empty()) {
+      parentStates.pop_back();
+    }
+  }
+
+  void onCharacterData(const XML_Char* data, const int len) {
+    if (!insideBody || nonVisibleDepth > 0 || len <= 0 || stopped) {
+      return;
+    }
+
+    const size_t codepointCount = countUtf8Codepoints(data, len);
+    if (codepointCount == 0) {
+      return;
+    }
+
+    // Start a new text node on first non-empty content after any element boundary.
+    // Only counting non-empty nodes matches KOReader's text()[N] indexing behavior,
+    // which skips empty text nodes created by bare <a id="anchor"/> anchors.
+    if (pendingTextNode) {
+      if (!textNodeIndexStack.empty()) {
+        textNodeIndexStack.back()++;
+      }
+      textNodeStartChars = visibleChars;
+      pendingTextNode = false;
+    }
+
+    const size_t nextVisibleChars = visibleChars + codepointCount;
+    if (targetVisibleChar <= nextVisibleChars) {
+      const size_t delta = targetVisibleChar - visibleChars;
+      const int texNode = textNodeIndexStack.empty() ? 0 : textNodeIndexStack.back();
+      const size_t charOff = visibleChars - textNodeStartChars + delta;
+      xpath = buildParagraphXPath(spineIndex, path, texNode, charOff);
+      stopped = true;
+      XML_StopParser(parser, XML_FALSE);
+      return;
+    }
+
+    visibleChars = nextVisibleChars;
+  }
+
+  XML_Parser parser = nullptr;
+  const size_t targetVisibleChar;
+  bool parseOk = true;
+  bool insideBody = false;
+  bool stopped = false;
+  bool pendingTextNode = true;
+  int depth = 0;
+  int bodyDepth = -1;
+  int paragraphDepth = 0;
+  int liDepth = 0;
+  int nonVisibleDepth = 0;
+  size_t visibleChars = 0;
+  size_t textNodeStartChars = 0;
+  std::vector<int> textNodeIndexStack;
+  std::vector<ParentState> parentStates;
+  std::vector<PathSegment> path;
+  std::string xpath;
+};
+
+std::string findXPathForElement(const std::shared_ptr<Epub>& epub, const int spineIndex, const uint16_t elementIndex,
+                                const char* tagName) {
+  if (!epub || elementIndex == 0 || spineIndex < 0 || spineIndex >= epub->getSpineItemsCount()) {
+    return "";
+  }
+
+  const auto href = epub->getSpineItem(spineIndex).href;
+  if (href.empty()) {
+    return "";
+  }
+
+  XPathElementResolver resolver(elementIndex, tagName);
+  if (!resolver.ok()) {
+    return "";
+  }
+
+  resolver.spineIndex = spineIndex;
+  if (!epub->readItemContentsToStream(href, resolver, 1024) || !resolver.finish()) {
+    return "";
+  }
+
+  if (resolver.hasMatch()) {
+    LOG_DBG("KOX", "Resolved %s %u in spine %d -> %s", tagName, elementIndex, spineIndex, resolver.getXPath().c_str());
+    return resolver.getXPath();
+  }
+
+  LOG_DBG("KOX", "%s %u not found in spine %d", tagName, elementIndex, spineIndex);
+  return "";
+}
+}  // namespace
+
+std::string ChapterXPathResolver::findXPathForParagraph(const std::shared_ptr<Epub>& epub, const int spineIndex,
+                                                        const uint16_t paragraphIndex) {
+  return findXPathForElement(epub, spineIndex, paragraphIndex, "p");
+}
+
+std::string ChapterXPathResolver::findXPathForListItem(const std::shared_ptr<Epub>& epub, const int spineIndex,
+                                                       const uint16_t listItemIndex) {
+  return findXPathForElement(epub, spineIndex, listItemIndex, "li");
+}
+
+std::string ChapterXPathResolver::findXPathForProgress(const std::shared_ptr<Epub>& epub, const int spineIndex,
+                                                       const float intraSpineProgress) {
+  if (!epub || spineIndex < 0 || spineIndex >= epub->getSpineItemsCount()) {
+    return "";
+  }
+
+  const auto href = epub->getSpineItem(spineIndex).href;
+  if (href.empty()) {
+    return "";
+  }
+
+  if (!(intraSpineProgress > 0.0f)) {
+    return "/body/DocFragment[" + std::to_string(spineIndex + 1) + "]/body";
+  }
+
+  ParagraphTextCounter counter;
+  if (!counter.ok() || !epub->readItemContentsToStream(href, counter, 1024) || !counter.finish()) {
+    return "";
+  }
+
+  const size_t totalVisibleChars = counter.totalVisibleChars();
+  if (totalVisibleChars == 0) {
+    return "";
+  }
+
+  const float clamped = std::max(0.0f, std::min(1.0f, intraSpineProgress));
+  const size_t targetVisibleChar =
+      std::max<size_t>(1, std::min(totalVisibleChars, static_cast<size_t>(std::ceil(clamped * totalVisibleChars))));
+
+  XPathProgressResolver resolver(targetVisibleChar);
+  if (!resolver.ok()) {
+    return "";
+  }
+
+  resolver.spineIndex = spineIndex;
+  if (!epub->readItemContentsToStream(href, resolver, 1024) || !resolver.finish()) {
+    return "";
+  }
+
+  if (resolver.hasMatch()) {
+    LOG_DBG("KOX", "Resolved progress %.3f in spine %d -> %s", intraSpineProgress, spineIndex,
+            resolver.getXPath().c_str());
+    return resolver.getXPath();
+  }
+
+  LOG_DBG("KOX", "Could not resolve progress %.3f in spine %d", intraSpineProgress, spineIndex);
+  return "";
+}
+
+std::string ChapterXPathResolver::findXPathForVisibleTextOffset(const std::shared_ptr<Epub>& epub, const int spineIndex,
+                                                                const uint32_t visibleTextOffset) {
+  if (!epub || spineIndex < 0 || spineIndex >= epub->getSpineItemsCount()) {
+    return "";
+  }
+
+  const auto href = epub->getSpineItem(spineIndex).href;
+  if (href.empty()) {
+    return "";
+  }
+
+  // XPath text offsets are one-based, while the cache records the first
+  // visible codepoint on a page with a zero-based offset.
+  const size_t targetVisibleChar = static_cast<size_t>(visibleTextOffset) + 1;
+  XPathProgressResolver resolver(targetVisibleChar);
+  if (!resolver.ok()) {
+    return "";
+  }
+  resolver.spineIndex = spineIndex;
+  if (!epub->readItemContentsToStream(href, resolver, 1024) || !resolver.finish()) {
+    return "";
+  }
+  if (!resolver.hasMatch()) {
+    LOG_DBG("KOX", "Could not resolve visible offset %lu in spine %d", static_cast<unsigned long>(visibleTextOffset),
+            spineIndex);
+    return "";
+  }
+  return resolver.getXPath();
+}
