+++ lib/EpubPosition/ChapterXPathResolver.h
@@ -1,0 +1,42 @@
+#pragma once
+
+#include <Epub.h>
+
+#include <cstdint>
+#include <memory>
+#include <string>
+
+class ChapterXPathResolver {
+ public:
+  /**
+   * Resolve the Nth paragraph in a spine item to its real XHTML ancestry path.
+   *
+   * Returns a KOReader-compatible path like:
+   * /body/DocFragment[8]/body/div[2]/section[1]/p[4]
+   *
+   * An empty string means parsing failed or the paragraph index was not found.
+   */
+  static std::string findXPathForParagraph(const std::shared_ptr<Epub>& epub, int spineIndex, uint16_t paragraphIndex);
+
+  /**
+   * Resolve the Nth list item in a spine item to its real XHTML ancestry path.
+   *
+   * An empty string means parsing failed or the list item index was not found.
+   */
+  static std::string findXPathForListItem(const std::shared_ptr<Epub>& epub, int spineIndex, uint16_t listItemIndex);
+
+  /**
+   * Resolve intra-spine progress to a real XHTML ancestry path plus text offset.
+   *
+   * Returns a KOReader-compatible path like:
+   * /body/DocFragment[8]/body/div[2]/section[1]/p[4]/text().96
+   *
+   * An empty string means parsing failed or the location could not be resolved.
+   */
+  static std::string findXPathForProgress(const std::shared_ptr<Epub>& epub, int spineIndex, float intraSpineProgress);
+
+  // Resolve a zero-based visible Unicode codepoint offset. This is the stable
+  // page position stored in CrossInk's section cache, independent of layout.
+  static std::string findXPathForVisibleTextOffset(const std::shared_ptr<Epub>& epub, int spineIndex,
+                                                   uint32_t visibleTextOffset);
+};
