+++ lib/EpubPosition/ProgressMapper.cpp
