+++ lib/EpubPosition/ProgressMapper.h
